FLEX-34809 Flex Code Coverage Tool
This is a tool designed to be run from the command line as part of a build
to create code coverage reports.

The tool has three pieces:
1. A Flex client that listens to the executed lines in the running SWF and
sends them over a socket to a Java server.
2. A Java server that reads the executed lines and writes them to disk.
3. A reporting tool that analyzes the executed lines and creates a report.

In addition to the source, this commit contains command line scripts, a
Flash Builder project file for the Flex code and an Eclipse project file
for the Java code.